refactor: collapse theme context
