--- src/context/index.ts
@@ -1,2 +1,0 @@
-export { ThemeProvider } from './ThemeContext'
-export { useThemeContext } from './useThemeContext'
