+++ src/context/index.tsx
@@ -1,0 +1,32 @@
+/* eslint-disable react-refresh/only-export-components -- ponytail: this context is intentionally collapsed; split again only if Fast Refresh issues appear. */
+import { createContext, use, type ReactNode } from 'react'
+
+import { useTheme, type Theme } from '../hooks/useTheme'
+
+export interface ThemeContextValue {
+  theme: Theme
+  resolvedTheme: 'light' | 'dark'
+  setTheme: (theme: Theme) => void
+  toggleTheme: () => void
+  isDark: boolean
+}
+
+const ThemeContext = createContext<ThemeContextValue | undefined>(undefined)
+
+interface ThemeProviderProps {
+  children: ReactNode
+}
+
+export function ThemeProvider({ children }: ThemeProviderProps) {
+  const themeValue = useTheme()
+
+  return <ThemeContext value={themeValue}>{children}</ThemeContext>
+}
+
+export function useThemeContext() {
+  const context = use(ThemeContext)
+  if (context === undefined) {
+    throw new Error('useThemeContext must be used within a ThemeProvider')
+  }
+  return context
+}
